Test of noclip camera in Pedestrian LOD visualisation.
It's not great due to limitations of GLUT.

Shift detection is tied to whether char input is captialised or not, so need to up and down key to notice changes.

If travelling for more than about a second, there's a clear jumpyness. We had this issue to begin with and added keyboard buffering to remedy it, not sure why this still occurs.

If you move the mouse too fast it goes outside the window, I think this is an issue on SDL however it polls much faster so its much harder to achieve.