Add outline of RestrictedFlowGraph example
This should provide the whole state machine and general behaviour.
Static graph loading is required next, to complete the behaviour of the mode.
Followed by graph-based communication to signifcantly improve performance.

It may be better to switch to a two state model,
to seperate agents which are on edges and agents which are on vertices,
improving cohesion but decreasing raw parallelism.