Updated simulation xslt to handle step functions specified in layers.
This does NOT fix the .xsd schema to permit step functions in layers

It first attempts to match layer function to any agent functions, if this fails it attempts to match with any step functions, if this also fails it adds preprocessor #error macro to prevent compilation.

For backwards compatibility, any step functions not mentioned in layers are called at the end.

Only concern here is if someone has agent and step functions of the same name, it will only call the agent functions. This seems stupid however.

Relevant to issue #97