Seperated functions.c from sim/fgpu_kernals.cu
This enables a much smaller file to be rebuilt when changes to functions.c (now .cu) are made.

I did a quick batch build of all debug versions, one issue remains.

The sugarscape example uses an agent condition which references a define made in functions.c

This should either be moved to environment vars, or compiler config.

Also requires linux make files to be updated.

#41