Do not specify the lib dir - use system installed versions
This is currently a global solution, but if it works will become a linux only sollution

Rather than provide .so files and libs, instead we should suggest to users that they provide their own copies, which must be on LD_LIBRARY_PATH etc.

This simplifies things on our end, and seems to be the most maintainable solution, although it does move this task to the user, rather than the the developer(us).

I think this may be a sensible option, as GL related libs seem to be relatively coupled with OpenGL support of hardware etc.